+++ Documentacion/Desarrollo/ProductoFuncional/BackendSIMP/routers/usuariosRouter.js
@@ -1,4 +1,3 @@
-// plantillaProductoRouter.js
 const crypto = require('crypto');
 const express = require('express');
 const router = express.Router();
@@ -82,36 +81,6 @@ router.delete('/borrar/:id', (req, res) => {
         if (error) console.error(error.message);
 
         res.json(`Se eliminó correctamente el usuario`);
-    });
-});
-
-router.post('/login', (req, res) => {
-    const nuevoUsuario = {
-        Correo: req.body.Correo,
-        Clave: req.body.Clave
-    };
-console.log(nuevoUsuario);
-    let hash = crypto.createHash('md5');
-    hash.update(nuevoUsuario.Clave);
-    let hashMD5 = hash.digest('hex');
-    nuevoUsuario.Clave = hashMD5
-
-    const query = `SELECT * FROM usuario WHERE Correo = '${nuevoUsuario.Correo}' AND Clave = '${nuevoUsuario.Clave}';`
-
-    conexion.query(query, async (error, resultado) => {
-        if (error) return console.error(error.message);
-        let rol = resultado[0].Rol_IdRol;
-        console.log(rol);
-
-        const tokenSession = await tokenSign(resultado[0]);
-
-
-        if (resultado.length > 0) {
-            console.log("Chido");
-            res.json(tokenSession)
-        } else {
-            res.json(false)
-        }
     });
 });
